// frontend/src/app/models/count.ts
export class Count {
    id: number;
    id_employee: number;
    employee_rut: number;
    working_day: string; 
    countscol: any;
    date: any;
    id_quarter: number;
    quarter_name: string;
    id_season: number;
    season_name: string;
    id_farm: number;
    farm_name: string;
    row: number;
    plant: any;
    registered_in_batch: boolean;
    cant: number;
    obj: number;
    diff: string;
    id_user: number;
    create_at: Date;
    updated_at: Date;
}